standardize macros #207 #154
