+++ src/3bc.h
@@ -7,11 +7,14 @@
 #define VERSION_3BC_MINOR 1
 #define VERSION_3BC_PATCH 2
 
+/** environment check **/
+#include "3bc_detect.h"
+
 /** prepare **/
 #include "3bc_macros.h"
 
 /** libaries **/
-#if !defined(_3BC_MOS6502)
+#if !defined(TBC_NOT_MATH)
 #include <math.h>
 #endif
 #include <ctype.h>
@@ -22,20 +25,20 @@
 #include <stdlib.h>
 #include <string.h>
 #include <unistd.h>
-#if defined(_3BC_COMPUTER)
+#if defined(TBC_P_COMPUTER)
 #include <signal.h>
 #include <time.h>
 #endif
-#if defined(_3BC_PC_UNIX)
+#if defined(TBC_USE_POSIX)
 #include <termios.h>
 #endif
-#if defined(_3BC_PC_WINDOWS) || defined(_3BC_PC_1970)
+#if defined(TBC_USE_CONIO)
 #include <conio.h>
 #endif
-#if defined(_3BC_ARDUINO)
+#if defined(TBC_USE_ARDUINO)
 #include <Arduino.h>
 #endif
-#if defined(_3BC_NUTTX)
+#if defined(TBC_KERNEL_NUTTX)
 #include <nuttx/config.h>
 #endif
 
@@ -49,15 +52,11 @@
 /* opcodes and witchcraft **/
 #include "3bc_register.h"
 
-/* some forbidden secret gimmicks **/
-#include "3bc_optional_custom.h"
-#include "3bc_optional_interpreter.h"
-
 /** compatibility fixes **/
 #include "3bc_support.h"
 
 /** program code **/
-#if !defined(_3BC_SCU)
+#if !defined(TBC_SCU)
 #include "cpu_boolean.c"
 #include "cpu_common.c"
 #include "cpu_jump.c"
@@ -67,6 +66,7 @@
 #include "cpu_sleep.c"
 #include "cpu_string.c"
 #include "driver_accumulator.c"
+#include "driver_custom.c"
 #include "driver_gpio.c"
 #include "driver_interrupt.c"
 #include "driver_memory.c"
@@ -79,5 +79,10 @@
 #include "ds_memory_llrbt.c"
 #include "ds_procedure_lifo.c"
 #include "ds_program_fifo.c"
+#include "interpreter_parser.c"
+#include "interpreter_readln.c"
+#include "interpreter_syntax.c"
+#include "interpreter_ticket.c"
+#include "interpreter_tokens.c"
 #endif
 #endif
