+++ src/3bc_macros.h
@@ -1,90 +1,4 @@
 /**
- * MACROS ENVIRONMENT (DETECT PLATAFORM/ARCHITECTURE)
- */
-
-/** Atmel AVR **/
-#if defined(ARDUINO_ARCH_AVR)
-#define _3BC_AVR
-#elif defined(__AVR_ATmega8__)
-#define _3BC_AVR
-#elif defined(__AVR_ATtiny24__) || defined(__AVR_ATtiny25__)
-#define _3BC_AVR
-#elif defined(__AVR_ATtiny44__) || defined(__AVR_ATtiny45__)
-#define _3BC_AVR
-#elif defined(__AVR_ATtiny84__) || defined(__AVR_ATtiny85__)
-#define _3BC_AVR
-#elif defined(__AVR_ATmega328__) || defined(__AVR_ATmega328P__)
-#define _3BC_AVR
-#elif defined(__AVR_ATmega644__) || defined(__AVR_ATmega644A__)                \
-    || defined(__AVR_ATmega644P__) || defined(__AVR_ATmega644PA__)
-#define _3BC_AVR
-#elif defined(__AVR_ATmega1280__) || defined(__AVR_ATmega1284__)               \
-    || defined(__AVR_ATmega1284P__)
-#define _3BC_AVR
-#elif defined(__AVR_ATmega2560__)
-#define _3BC_AVR
-#endif
-
-#if defined(__CC65_STD__) || defined(__CC65_STD_CC65__)
-#define _3BC_MOS6502
-#endif
-
-#if defined(__nuttx__)
-#define _3BC_NUTTX
-#endif
-
-/**
- * PLATAFORM: ARDUINO
- *
- * NOTE: _3BC_SCU allows arduino IDE compilation in parts.
- * NOTE: _3BC_COMPACT reduces rom size.
- */
-#if defined(ARDUINO)
-#define _3BC_SCU
-#define _3BC_COMPACT
-#define _3BC_ARDUINO
-/**
- * PLATAFORM: OLD COMPUTER
- */
-#elif defined(_3BC_MOS6502)
-#define _3BC_COMPACT
-#define _3BC_COMPUTER
-#define _3BC_PC_1970
-#if !defined(_3BC_ENABLE_INTERPRETER) && !defined(_3BC_DISABLE_INTERPRETER)
-#define _3BC_DISABLE_INTERPRETER
-#endif
-#elif defined(_3BC_NUTTX)
-/**
- * PLATAFORM: NUTTX RTOS
- */
-#define _3BC_COMPUTER
-#if !defined(_3BC_ENABLE_INTERPRETER) && !defined(_3BC_DISABLE_INTERPRETER)
-#define _3BC_ENABLE_INTERPRETER
-#endif
-#define _3BC_PC_UNIX
-/**
- * PLATAFORM: MODERN COMPUTER
- */
-#else
-#define _3BC_COMPUTER
-#if !defined(_3BC_ENABLE_INTERPRETER) && !defined(_3BC_DISABLE_INTERPRETER)
-#define _3BC_ENABLE_INTERPRETER
-#endif
-#if defined(_WIN32)
-#define _3BC_PC_WINDOWS
-#else
-#define _3BC_PC_UNIX
-#endif
-#endif
-
-/**
- * PARTITIONED COMPILATION
- */
-#ifndef _3BC_SCU
-#define _3BC_SCU_FIX
-#endif
-
-/**
  * CONSTS MACROS
  */
 #define LLRBT_BLACK (false)
@@ -95,7 +9,7 @@
 #endif
 
 #ifndef LABEL_HASH_SIZE
-#ifdef _3BC_ARDUINO
+#ifdef TBC_OPT_COMPACT
 #define LABEL_HASH_SIZE (8)
 #else
 #define LABEL_HASH_SIZE (128)
@@ -261,4 +175,4 @@
     ;                                                                          \
     a->t.type = STREAM_TYPE_FUNCTION_CALL;                                     \
     a->t.io.lambda = l;
-#define lang_3bc_custom custom_3bc_func_set
+#define lang_3bc_custom driver_custom_set
