+++ src/3bc_register.h
@@ -98,7 +98,7 @@
  *
  * (required for avr compiler in arduino ide)
  */
-#ifdef _3BC_SCU_FIX
+#ifdef TBC_SOURCE_REGISTERS
 void instruction_3bc(struct app_3bc_s* const app, register_3bc_t reg,
     address_3bc_t address, data_3bc_t value)
 {
