+++ src/driver_custom.c
@@ -1,17 +1,15 @@
+#define TBC_SOURCE_ENTRY
 #include "3bc.h"
 
-#if !(defined(ARDUINO) && defined(_3BC_SCU_FIX_2))                             \
-    || (!defined(ARDUINO) && defined(_3BC_SCU_FIX))
-
-#if !defined(_3BC_ENABLE_CUSTOM)
-void custom_3bc_func_call(struct app_3bc_s* const app, register_3bc_t reg,
+#if !defined(TBC_SCU_OPTIONAL_FIX)
+#if !defined(TBC_CUSTOM)
+void driver_custom_call(struct app_3bc_s* const app, register_3bc_t reg,
     address_3bc_t address, data_3bc_t value)
 {
     driver_program_error(app, ERROR_CPU_RESERVED);
 }
 
 #else
-
 function_3bc_t* custom_funcs = NULL;
 unsigned char last_func = 0;
 
@@ -19,7 +17,7 @@ unsigned char last_func = 0;
  * DS: Dynamic Array
  *
  */
-void custom_3bc_func_set(struct app_3bc_s* const app, cpumode_3bc_t cpu_mode,
+void driver_custom_set(struct app_3bc_s* const app, cpumode_3bc_t cpu_mode,
     register_3bc_t reg, function_3bc_t lambda)
 {
     unsigned char atual_func = ((cpu_mode / 10) - 1) * 6 + reg;
@@ -45,7 +43,7 @@ void custom_3bc_func_set(struct app_3bc_s* const app, cpumode_3bc_t cpu_mode,
     custom_funcs[atual_func] = lambda;
 }
 
-void custom_3bc_func_call(struct app_3bc_s* const app, register_3bc_t reg,
+void driver_custom_call(struct app_3bc_s* const app, register_3bc_t reg,
     address_3bc_t address, data_3bc_t value)
 {
     unsigned char atual_func = ((app->cpu_mode / 10) - 1) * 6 + reg;
