+++ src/driver_tty.c
@@ -1,14 +1,14 @@
-#define _3BC_SCU_FIX_2
+#define TBC_SOURCE_ENTRY
 #include "3bc.h"
 
-#if defined(_3BC_PC_UNIX)
+#if defined(TBC_USE_POSIX)
 struct termios term_old_attr;
 struct termios term_new_attr;
 #endif
 
 void driver_tty_init()
 {
-#if defined(_3BC_PC_UNIX)
+#if defined(TBC_USE_POSIX)
     /**
      * Turn possible terminal uncannonical mode
      * without conio.h in linux/unix builds
@@ -26,13 +26,13 @@ void driver_tty_init()
 
 void driver_tty_exit()
 {
-#if defined(_3BC_COMPUTER)
+#if defined(TBC_P_COMPUTER)
     /** clear buffers **/
     fflush(stderr);
     fflush(stdout);
 #endif
 
-#if defined(_3BC_PC_UNIX)
+#if defined(TBC_USE_POSIX)
     /** reset terminal to default mode (linux/unix) **/
     tcsetattr(STDIN_FILENO, TCSANOW, &term_old_attr);
 #endif
@@ -54,13 +54,13 @@ data_3bc_t driver_tty_input(
         invalid = false;
 
 /** capture input **/
-#if defined(_3BC_PC_1970)
+#if defined(TBC_P_COMPUTER_OLD)
         c[0] = cgetc();
-#elif defined(_3BC_PC_UNIX)
+#elif defined(TBC_USE_POSIX)
         tcsetattr(STDIN_FILENO, TCSANOW, &term_new_attr);
         c[0] = getchar();
         tcsetattr(STDIN_FILENO, TCSANOW, &term_old_attr);
-#elif defined(_3BC_PC_WINDOWS)
+#elif defined(TBC_USE_CONIO)
         /** exclusive function of the conio.h library **/
         c[0] = getch();
 #endif
@@ -181,7 +181,7 @@ void driver_tty_output_raw(
         return;
     }
 #endif
-#if defined(_3BC_COMPUTER)
+#if defined(TBC_P_COMPUTER)
     /** stream standard c output **/
     if (tty.type == STREAM_TYPE_COMPUTER_STD) {
         fputs(string, tty.io.stream);
