+++ src/ds_hypervisor_darray.c
@@ -1,4 +1,4 @@
-#define _3BC_SCU_FIX_2
+#define TBC_SOURCE_ENTRY
 #include "3bc.h"
 
 static app_3bc_id machines_count;
